requirements.txt
fastapi
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pycryptodome
pydantic
motor
bson
uvicorn